cmd/blsync,beacon: move blsync code into beacon
