+++ beacon/beaclient/client.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-package main
+package beaclient
 
 import (
 	"context"
@@ -27,6 +27,7 @@ import (
 	eth2api "github.com/attestantio/go-eth2-client/api"
 	eth2http "github.com/attestantio/go-eth2-client/http"
 	eth2spec "github.com/attestantio/go-eth2-client/spec"
+	"github.com/ethereum/go-ethereum/beacon/types"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/rs/zerolog"
 )
@@ -38,13 +39,13 @@ var (
 	beaconLightClientFinalityUpdate   = "eth/v1/beacon/light_client/finality_update"
 )
 
-type BeaconClient struct {
+type Client struct {
 	ctx    context.Context
 	url    string
 	client eth2client.Service
 }
 
-func NewBeaconClient(ctx context.Context, server string) (*BeaconClient, error) {
+func NewClient(ctx context.Context, server string) (*Client, error) {
 	client, err := eth2http.New(
 		ctx,
 		eth2http.WithAddress(server),
@@ -54,14 +55,14 @@ func NewBeaconClient(ctx context.Context, server string) (*BeaconClient, error) 
 	if err != nil {
 		return nil, err
 	}
-	return &BeaconClient{
+	return &Client{
 		ctx:    ctx,
 		url:    server,
 		client: client,
 	}, nil
 }
 
-func (c *BeaconClient) Bootstrap(root common.Hash) (*Bootstrap, error) {
+func (c *Client) Bootstrap(root common.Hash) (*types.Bootstrap, error) {
 	resp, err := http.Get(fmt.Sprintf("%s/%s/%s", c.url, beaconLightClientBootstrap, root.String()))
 	if err != nil {
 		return nil, fmt.Errorf("failed http request: %w", err)
@@ -70,14 +71,14 @@ func (c *BeaconClient) Bootstrap(root common.Hash) (*Bootstrap, error) {
 	if err != nil {
 		return nil, fmt.Errorf("failed to read response: %w", err)
 	}
-	var bs Bootstrap
+	var bs types.Bootstrap
 	if err := json.Unmarshal(b, &bs); err != nil {
 		return nil, fmt.Errorf("failed to unmarshal data: %v", err)
 	}
 	return &bs, nil
 }
 
-func (c *BeaconClient) GetRangeUpdate(start, count int) ([]*LightClientUpdate, error) {
+func (c *Client) GetRangeUpdate(start, count int) ([]*types.LightClientUpdate, error) {
 	resp, err := http.Get(fmt.Sprintf("%s/%s?start_period=%d&count=%d", c.url, beaconLightClientUpdate, start, count))
 	if err != nil {
 		return nil, fmt.Errorf("failed http request: %w", err)
@@ -86,7 +87,7 @@ func (c *BeaconClient) GetRangeUpdate(start, count int) ([]*LightClientUpdate, e
 	if err != nil {
 		return nil, fmt.Errorf("failed to read response: %w", err)
 	}
-	var u []*LightClientUpdate
+	var u []*types.LightClientUpdate
 	if err := json.Unmarshal(b, &u); err != nil {
 		return nil, fmt.Errorf("failed to unmarshal data: %v", err)
 	}
@@ -94,7 +95,7 @@ func (c *BeaconClient) GetRangeUpdate(start, count int) ([]*LightClientUpdate, e
 
 }
 
-func (c *BeaconClient) GetOptimisticUpdate() (*LightClientUpdate, error) {
+func (c *Client) GetOptimisticUpdate() (*types.LightClientUpdate, error) {
 	resp, err := http.Get(fmt.Sprintf("%s/%s", c.url, beaconLightClientOptimisticUpdate))
 	if err != nil {
 		return nil, fmt.Errorf("failed http request: %w", err)
@@ -103,14 +104,14 @@ func (c *BeaconClient) GetOptimisticUpdate() (*LightClientUpdate, error) {
 	if err != nil {
 		return nil, fmt.Errorf("failed to read response: %w", err)
 	}
-	var u LightClientUpdate
+	var u types.LightClientUpdate
 	if err := json.Unmarshal(b, &u); err != nil {
 		return nil, fmt.Errorf("failed to unmarshal data: %v", err)
 	}
 	return &u, nil
 }
 
-func (c *BeaconClient) GetFinalityUpdate() (*LightClientUpdate, error) {
+func (c *Client) GetFinalityUpdate() (*types.LightClientUpdate, error) {
 	resp, err := http.Get(fmt.Sprintf("%s/%s", c.url, beaconLightClientFinalityUpdate))
 	if err != nil {
 		return nil, fmt.Errorf("failed http request: %w", err)
@@ -119,7 +120,7 @@ func (c *BeaconClient) GetFinalityUpdate() (*LightClientUpdate, error) {
 	if err != nil {
 		return nil, fmt.Errorf("failed to read response: %w", err)
 	}
-	var u LightClientUpdate
+	var u types.LightClientUpdate
 	if err := json.Unmarshal(b, &u); err != nil {
 		return nil, fmt.Errorf("failed to unmarshal data: %v", err)
 	}
@@ -127,7 +128,7 @@ func (c *BeaconClient) GetFinalityUpdate() (*LightClientUpdate, error) {
 
 }
 
-func (c *BeaconClient) GetBlock(root common.Hash) (*eth2spec.VersionedSignedBeaconBlock, error) {
+func (c *Client) GetBlock(root common.Hash) (*eth2spec.VersionedSignedBeaconBlock, error) {
 	provider, ok := c.client.(eth2client.SignedBeaconBlockProvider)
 	if !ok {
 		return nil, fmt.Errorf("beacon server does not support retrieving blocks")
